Faster_R-CNN/requirements.txt
matplotlib==3.1.3
numpy==1.19.0
numpydoc==0.9.2
opencv-python==4.1.0.25
pandas==1.0.3
Pillow==7.1.2
torch==1.7.0+cpu
torchaudio==0.7.0
torchvision==0.8.1+cpu